refactor: organize client files by mvc
