--- src/client/components/codeEditor.css
@@ -1,34 +1,0 @@
-.cm-editor {
-  background: rgba(0, 0, 0, 0);
-  transition: background-color 0.3s ease;
-  color: #cccccc;
-  font-size: 18px;
-  display: flex;
-  height: 100%;
-  padding: 0.5em;
-}
-
-.cm-line {
-  background-color: rgba(0, 0, 0, 0.6);
-  transition: background-color 0.3s ease;
-  color: #cccccc;
-  width: fit-content;
-  padding-left: 0 !important;
-  margin-bottom: 0.25em;
-}
-
-.cm-line span:first-child {
-  padding-left: 6px;
-}
-
-.cm-activeLine {
-  background-color: rgba(0, 0, 0, 0.9) !important;
-}
-
-.em-editor .cm-cursor {
-  border: 1px solid red;
-}
-
-.ͼ1.cm-focused {
-  outline: 0vh;
-}
